+++ src/configWoosmapGeofencingAndroid.ts
@@ -5,7 +5,6 @@ import {
 } from "@expo/config-plugins";
 
 import { ConfigProps } from "./types";
-import { addMetaDataItemToMainApplication } from "@expo/config-plugins/build/android/Manifest";
 
 const withSDKAndroidManifest: ConfigPlugin<ConfigProps> = (config, props) => {
   config = withAndroidManifest(config, (config) => {
@@ -13,16 +12,29 @@ const withSDKAndroidManifest: ConfigPlugin<ConfigProps> = (config, props) => {
       config.modResults,
     );
 
+    var {
+      useAndroidBackgroundLocation,
+      useAndroidBluetooth,
+    } = props;
+   useAndroidBackgroundLocation =  useAndroidBackgroundLocation == undefined? true: useAndroidBackgroundLocation; //Default true
+   useAndroidBluetooth =  useAndroidBluetooth == undefined? true: useAndroidBluetooth; //Default true
+    
     // Add location and BLE permissions to the <manifest> tag
     const permissions = [
       "android.permission.ACCESS_COARSE_LOCATION",
       "android.permission.ACCESS_FINE_LOCATION",
-      "android.permission.ACCESS_BACKGROUND_LOCATION",
-      "android.permission.BLUETOOTH",
-      "android.permission.BLUETOOTH_ADMIN",
-      "android.permission.BLUETOOTH_SCAN",
     ];
-
+   
+    if(useAndroidBackgroundLocation == true){
+      permissions.push("android.permission.ACCESS_BACKGROUND_LOCATION")
+    }
+    
+    if(useAndroidBluetooth == true){
+      permissions.push("android.permission.BLUETOOTH")
+      permissions.push("android.permission.BLUETOOTH_ADMIN")
+      permissions.push("android.permission.BLUETOOTH_SCAN")
+    }
+    
     // Ensure each permission is added only once
     permissions.forEach((permission) => {
       if (
@@ -45,8 +57,5 @@ const withSDKAndroidManifest: ConfigPlugin<ConfigProps> = (config, props) => {
 
 export const withAndroidSdk: ConfigPlugin<ConfigProps> = (config, props) => {
   config = withSDKAndroidManifest(config, props);
-  //   config = withSDKEntitlements(config, props);
-  //   config = withSDKXcodeProject(config, props);
-  //   config = withSDKDangerousMod(config, props);
   return config;
 };