+++ src/types.ts
@@ -2,4 +2,6 @@ export type ConfigProps = {
   locationAlwaysAndWhenInUsePermission: string;
   locationAlwaysPermission: string;
   locationWhenInUsePermission: string;
+  useAndroidBackgroundLocation: boolean;
+  useAndroidBluetooth:boolean;
 };